Deploiement/requirements.txt
flask==2.2.3
werkzeug==2.2.3
pandas==1.5.3
numpy==1.24.2
scikit-learn==1.2.2
xgboost==1.7.5
joblib==1.2.0
gunicorn==20.1.0
prometheus-client==0.16.0
flask-prometheus-metrics==1.0.0
Flask-MonitoringDashboard==3.1.1
python-dotenv==1.0.0